Update Linux instructions.
It seems to me that the instructions for installing on Linux are outdated. Following it exactly, I was unable to install it in Lubuntu.

First I had some problems with the JDK that I solved by updating to 11.

And then the git url doesn't seem to work anymore because it didn't connect. I updated it to the repository url and it worked.